+++ src/core/ref.jl
@@ -31,6 +31,20 @@ function _get_function_from_head_curve(head_curve::Array{Tuple{Float64,Float64}}
 end
 
 
+function calc_demand_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
+    # Initialize the dictionaries for minimum and maximum heads.
+    demand_min = Dict((i, -Inf) for (i, junc) in ref(wm, n, :dispatchable_junction))
+    demand_max = Dict((i, Inf) for (i, junc) in ref(wm, n, :dispatchable_junction))
+
+    for (i, junction) in ref(wm, n, :dispatchable_junction)
+        demand_min[i] = max(demand_min[i], junction["demand_min"])
+        demand_max[i] = min(demand_max[i], junction["demand_max"])
+    end
+
+    return demand_min, demand_max
+end
+
+
 function calc_head_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
     # Compute the maximum elevation of all nodes in the network.
     max_head = maximum(node["elevation"] for (i, node) in ref(wm, n, :node))
@@ -52,16 +66,6 @@ function calc_head_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
     head_max = Dict((i, max_head) for (i, node) in ref(wm, n, :node))
 
     for (i, node) in ref(wm, n, :node)
-        # Override the minimum head value at node i, if additional data is present.
-        if haskey(node, "minimumHead")
-            head_min[i] = max(head_min[i], node["minimumHead"])
-        end
-
-        # Override the maximum head value at node i, if additional data is present.
-        if haskey(node, "maximumHead")
-            head_max[i] = min(head_max[i], node["maximumHead"])
-        end
-
         num_junctions = length(ref(wm, n, :node_junction, i))
         num_reservoirs = length(ref(wm, n, :node_reservoir, i))
         num_tanks = length(ref(wm, n, :node_tank, i))
@@ -75,8 +79,14 @@ function calc_head_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
     for (i, reservoir) in ref(wm, n, :reservoir)
         # Fix head values at nodes with reservoirs to predefined values.
         node_id = reservoir["node"]
-        head_min[node_id] = ref(wm, n, :node, node_id)["head"]
-        head_max[node_id] = ref(wm, n, :node, node_id)["head"]
+
+        if !reservoir["dispatchable"]
+            head_min[node_id] = ref(wm, n, :node, node_id)["head"]
+            head_max[node_id] = ref(wm, n, :node, node_id)["head"]
+        else
+            head_min[node_id] = ref(wm, n, :node, node_id)["h_min"]
+            head_max[node_id] = ref(wm, n, :node, node_id)["h_max"]
+        end
     end
 
     for (i, tank) in ref(wm, n, :tank)
@@ -93,6 +103,11 @@ function calc_head_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
         head_min[node_to] = min(head_min[node_to], h_setting)
     end
 
+    for (i, node) in ref(wm, n, :node)
+        haskey(node, "h_min") && (head_min[i] = max(head_min[i], node["h_min"]))
+        haskey(node, "h_max") && (head_max[i] = min(head_max[i], node["h_max"]))
+    end
+
     # Return the dictionaries of lower and upper bounds.
     return head_min, head_max
 end
@@ -102,8 +117,10 @@ function calc_flow_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
     h_lb, h_ub = calc_head_bounds(wm, n)
     alpha = ref(wm, n, :alpha)
 
-    junctions = values(ref(wm, n, :junction))
-    sum_demand = sum(abs(junction["demand"]) for junction in junctions)
+    nondispatchable_junctions = values(ref(wm, n, :nondispatchable_junction))
+    sum_demand = length(nondispatchable_junctions) > 0 ? sum(junc["demand"] for junc in nondispatchable_junctions) : 0.0
+    dispatchable_junctions = values(ref(wm, n, :dispatchable_junction))
+    sum_demand += length(dispatchable_junctions) > 0 ? sum(junc["demand_max"] for junc in dispatchable_junctions) : 0.0
 
     if :time_step in keys(ref(wm, n))
         time_step = ref(wm, n, :time_step)
@@ -144,6 +161,9 @@ function calc_flow_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
                     ub[name][a][r_id] = min(ub[name][a][r_id], rate_bound)
                 end
             end
+
+            haskey(comp, "q_min") && (lb[name][a][1] = max(lb[name][a][1], comp["q_min"]))
+            haskey(comp, "q_max") && (ub[name][a][1] = min(ub[name][a][1], comp["q_max"]))
         end
     end
 
@@ -151,8 +171,10 @@ function calc_flow_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
     ub["pressure_reducing_valve"] = Dict{Int,Array{Float64}}()
 
     for (a, prv) in ref(wm, n, :pressure_reducing_valve)
-        lb["pressure_reducing_valve"][a] = [0.0]
-        ub["pressure_reducing_valve"][a] = [sum_demand]
+        name = "pressure_reducing_valve"
+        lb[name][a], ub[name][a] = [0.0], [sum_demand]
+        haskey(prv, "q_min") && (lb[name][a][1] = max(lb[name][a][1], prv["q_min"]))
+        haskey(prv, "q_max") && (ub[name][a][1] = min(ub[name][a][1], prv["q_max"]))
     end
 
     lb["pump"], ub["pump"] = Dict{Int,Array{Float64}}(), Dict{Int,Array{Float64}}()
@@ -162,7 +184,9 @@ function calc_flow_bounds(wm::AbstractWaterModel, n::Int=wm.cnw)
         c = _get_function_from_head_curve(head_curve)
         q_max = (-c[2] + sqrt(c[2]^2 - 4.0*c[1]*c[3])) * inv(2.0*c[1])
         q_max = max(q_max, (-c[2] - sqrt(c[2]^2 - 4.0*c[1]*c[3])) * inv(2.0*c[1]))
-        lb["pump"][a], ub["pump"][a] = [[0.0], [q_max]]
+        lb["pump"][a], ub["pump"][a] = [[0.0], [min(sum_demand, q_max)]]
+        haskey(pump, "q_min") && (lb["pump"][a][1] = max(lb["pump"][a][1], pump["q_min"]))
+        haskey(pump, "q_max") && (ub["pump"][a][1] = min(ub["pump"][a][1], pump["q_max"]))
     end
 
     return lb, ub