+++ src/form/milpr.jl
@@ -13,7 +13,7 @@ function _get_head_loss_oa(q::JuMP.VariableRef, q_hat::Float64, alpha::Float64)
 end
 
 
-function _get_head_loss_cv_oa(q::JuMP.VariableRef, z::Union{JuMP.VariableRef, JuMP.GenericAffExpr}, q_hat::Float64, alpha::Float64)
+function _get_head_loss_oa_z(q::JuMP.VariableRef, z::Union{JuMP.VariableRef, JuMP.GenericAffExpr}, q_hat::Float64, alpha::Float64)
     return q_hat^alpha*z + alpha * q_hat^(alpha - 1.0) * (q - q_hat*z)
 end
 
@@ -42,7 +42,7 @@ function constraint_pump_head_gain(wm::AbstractMILPRModel, n::Int, a::Int, node_
     c_4 = JuMP.@constraint(wm.model, lambda[a, end] <= x_pw[a, end])
 
     # Add a constraint for the flow piecewise approximation.
-    breakpoints = range(_q_eps, stop=JuMP.upper_bound(qp), length=pump_breakpoints)
+    breakpoints = range(0.0, stop=JuMP.upper_bound(qp), length=pump_breakpoints)
     qp_lhs = sum(breakpoints[k] * lambda[a, k] for k in 1:pump_breakpoints)
     c_5 = JuMP.@constraint(wm.model, qp_lhs == qp)
 
@@ -99,12 +99,15 @@ function constraint_pipe_head_loss(wm::AbstractMILPRModel, n::Int, a::Int, node_
     pipe_breakpoints = get(wm.ext, :pipe_breakpoints, 0)
     if pipe_breakpoints <= 0 return end
 
+    # Get the variable denoting flow directionality.
+    y = var(wm, n, :y_pipe, a)
+
     # Add constraints corresponding to positive outer-approximations.
     qp, dhp = var(wm, n, :qp_pipe, a), var(wm, n, :dhp_pipe, a)
 
     for qp_hat in range(0.0, stop=JuMP.upper_bound(qp), length=pipe_breakpoints)
-        lhs = r * _get_head_loss_oa(qp, qp_hat, alpha)
-        c_p = JuMP.@constraint(wm.model, lhs <= inv(L) * dhp)
+        lhs = _get_head_loss_oa_z(qp, y, qp_hat, alpha)
+        c_p = JuMP.@constraint(wm.model, r * lhs <= inv(L) * dhp)
         append!(con(wm, n, :head_loss)[a], [c_p])
     end
 
@@ -112,8 +115,8 @@ function constraint_pipe_head_loss(wm::AbstractMILPRModel, n::Int, a::Int, node_
     qn, dhn = var(wm, n, :qn_pipe, a), var(wm, n, :dhn_pipe, a)
 
     for qn_hat in range(0.0, stop=JuMP.upper_bound(qn), length=pipe_breakpoints)
-        lhs = r * _get_head_loss_oa(qn, qn_hat, alpha)
-        c_n = JuMP.@constraint(wm.model, lhs <= inv(L) * dhn)
+        lhs = _get_head_loss_oa_z(qn, 1.0 - y, qn_hat, alpha)
+        c_n = JuMP.@constraint(wm.model, r * lhs <= inv(L) * dhn)
         append!(con(wm, n, :head_loss)[a], [c_n])
     end
 end
@@ -125,14 +128,16 @@ function constraint_check_valve_head_loss(wm::AbstractMILPRModel, n::Int, a::Int
     if pipe_breakpoints <= 0 return end
 
     # Get common variables for outer approximation constraints.
-    qp, z = var(wm, n, :qp_check_valve, a), var(wm, n, :z_check_valve, a)
-    dhp = var(wm, n, :dhp_check_valve, a)
+    y, z = var(wm, n, :y_check_valve, a), var(wm, n, :z_check_valve, a)
+    qp, dhp = var(wm, n, :qp_check_valve, a), var(wm, n, :dhp_check_valve, a)
 
     # Add outer approximation constraints.
     for qp_hat in range(0.0, stop=JuMP.upper_bound(qp), length=pipe_breakpoints)
-        lhs = _get_head_loss_cv_oa(qp, z, qp_hat, ref(wm, n, :alpha))
-        c_p = JuMP.@constraint(wm.model, r * lhs <= inv(L) * dhp)
-        append!(con(wm, n, :head_loss)[a], [c_p])
+        lhs_1 = _get_head_loss_oa_z(qp, y, qp_hat, ref(wm, n, :alpha))
+        c_p_1 = JuMP.@constraint(wm.model, r * lhs_1 <= inv(L) * dhp)
+        lhs_2 = _get_head_loss_oa_z(qp, z, qp_hat, ref(wm, n, :alpha))
+        c_p_2 = JuMP.@constraint(wm.model, r * lhs_2 <= inv(L) * dhp)
+        append!(con(wm, n, :head_loss)[a], [c_p_1, c_p_2])
     end
 end
 
@@ -145,20 +150,24 @@ function constraint_shutoff_valve_head_loss(wm::AbstractMILPRModel, n::Int, a::I
     # Get common data for outer approximation constraints.
     qp, qn = var(wm, n, :qp_shutoff_valve, a), var(wm, n, :qn_shutoff_valve, a)
     dhp, dhn = var(wm, n, :dhp_shutoff_valve, a), var(wm, n, :dhn_shutoff_valve, a)
-    y = var(wm, n, :y_shutoff_valve, a)
+    y, z = var(wm, n, :y_shutoff_valve, a), var(wm, n, :z_shutoff_valve, a)
 
     # Add outer approximation constraints for positively-directed flow.
     for qp_hat in range(0.0, stop=JuMP.upper_bound(qp), length=pipe_breakpoints)
-        lhs = _get_head_loss_cv_oa(qp, y, qp_hat, ref(wm, n, :alpha))
-        c_p = JuMP.@constraint(wm.model, L * r * lhs <= dhp)
-        append!(con(wm, n, :head_loss)[a], [c_p])
+        lhs_1 = _get_head_loss_oa_z(qp, y, qp_hat, ref(wm, n, :alpha))
+        c_p_1 = JuMP.@constraint(wm.model, L * r * lhs_1 <= dhp)
+        lhs_2 = _get_head_loss_oa_z(qp, z, qp_hat, ref(wm, n, :alpha))
+        c_p_2 = JuMP.@constraint(wm.model, L * r * lhs_2 <= dhp)
+        append!(con(wm, n, :head_loss)[a], [c_p_1, c_p_2])
     end
 
     # Add outer approximation constraints for negatively-directed flow.
     for qn_hat in range(0.0, stop=JuMP.upper_bound(qn), length=pipe_breakpoints)
-        lhs = _get_head_loss_cv_oa(qn, (1.0 - y), qn_hat, ref(wm, n, :alpha))
-        c_n = JuMP.@constraint(wm.model, L * r * lhs <= dhn)
-        append!(con(wm, n, :head_loss)[a], [c_n])
+        lhs_1 = _get_head_loss_oa_z(qn, 1.0 - y, qn_hat, ref(wm, n, :alpha))
+        c_n_1 = JuMP.@constraint(wm.model, L * r * lhs_1 <= dhn)
+        lhs_2 = _get_head_loss_oa_z(qn, z, qn_hat, ref(wm, n, :alpha))
+        c_n_2 = JuMP.@constraint(wm.model, L * r * lhs_2 <= dhn)
+        append!(con(wm, n, :head_loss)[a], [c_n_1, c_n_2])
     end
 end
 
@@ -192,7 +201,7 @@ function objective_owf(wm::AbstractMILPRModel)
                 qp_ub = JuMP.upper_bound(qp)
 
                 # Generate a set of uniform flow and cubic function breakpoints.
-                breakpoints = range(_q_eps, stop=qp_ub, length=pump_breakpoints)
+                breakpoints = range(0.0, stop=qp_ub, length=pump_breakpoints)
                 f = _calc_cubic_flow_values(collect(breakpoints), curve_fun)
 
                 # Get pump efficiency data.