+++ src/form/undirected_flow.jl
@@ -97,23 +97,22 @@ function constraint_check_valve_common(wm::AbstractUndirectedModel, n::Int, a::I
     # Get flow and check valve status variables.
     q, z = var(wm, n, :q_check_valve, a), var(wm, n, :z_check_valve, a)
 
-    # If the check valve is open, flow must be appreciably nonnegative.
+    # If the check valve is closed, flow must be zero.
     c_1 = JuMP.@constraint(wm.model, q <= JuMP.upper_bound(q) * z)
-    c_2 = JuMP.@constraint(wm.model, q >= _q_eps * z)
 
     # Get head variables for from and to nodes.
     h_i, h_j = var(wm, n, :h, node_fr), var(wm, n, :h, node_to)
 
     # When the check valve is open, negative head loss is not possible.
     dh_lb = JuMP.lower_bound(h_i) - JuMP.upper_bound(h_j)
-    c_3 = JuMP.@constraint(wm.model, h_i - h_j >= (1.0 - z) * dh_lb)
+    c_2 = JuMP.@constraint(wm.model, h_i - h_j >= (1.0 - z) * min(0.0, dh_lb))
 
     # When the check valve is closed, positive head loss is not possible.
     dh_ub = JuMP.upper_bound(h_i) - JuMP.lower_bound(h_j)
-    c_4 = JuMP.@constraint(wm.model, h_i - h_j <= z * dh_ub)
+    c_3 = JuMP.@constraint(wm.model, h_i - h_j <= z * dh_ub)
 
     # Append the constraint array.
-    append!(con(wm, n, :check_valve, a), [c_1, c_2, c_3, c_4])
+    append!(con(wm, n, :check_valve, a), [c_1, c_2, c_3])
 end
 
 
@@ -168,9 +167,9 @@ function constraint_pump_common(wm::AbstractUndirectedModel, n::Int, a::Int, nod
     c_2 = JuMP.@constraint(wm.model, q >= _q_eps * z)
 
     # If the pump is off, decouple the head difference relationship.
-    dhn_lb = JuMP.lower_bound(h_j) - JuMP.upper_bound(h_i)
+    dhn_lb = min(0.0, JuMP.lower_bound(h_j) - JuMP.upper_bound(h_i))
     c_3 = JuMP.@constraint(wm.model, h_j - h_i - g >= dhn_lb * (1.0 - z))
-    dhn_ub = JuMP.upper_bound(h_j) - JuMP.lower_bound(h_i)
+    dhn_ub = max(0.0, JuMP.upper_bound(h_j) - JuMP.lower_bound(h_i))
     c_4 = JuMP.@constraint(wm.model, h_j - h_i - g <= dhn_ub * (1.0 - z))
 
     # Append the constraint array.
@@ -178,6 +177,15 @@ function constraint_pump_common(wm::AbstractUndirectedModel, n::Int, a::Int, nod
 end
 
 function constraint_pipe_common(wm::AbstractUndirectedModel, n::Int, a::Int, node_fr::Int, node_to::Int, alpha::Float64, L::Float64, r::Float64)
+    # For undirected formulations, there are no constraints, here.
+end
+
+function constraint_node_directionality(
+    wm::AbstractWaterModel, n::Int, i::Int, check_valve_fr::Array{Int},
+    check_valve_to::Array{Int}, pipe_fr::Array{Int}, pipe_to::Array{Int},
+    pump_fr::Array{Int}, pump_to::Array{Int}, pressure_reducing_valve_fr::Array{Int},
+    pressure_reducing_valve_to::Array{Int}, shutoff_valve_fr::Array{Int},
+    shutoff_valve_to::Array{Int})
     # For undirected formulations, there are no constraints, here.
 end