+++ test/runtests.jl
@@ -44,6 +44,8 @@ include("common.jl")
 
     include("des.jl")
 
+    include("util.jl")
+
     include("warm_start.jl")
 
 end